+++ example/lib/pages/simple/photo_view_demo.dart
@@ -1,5 +1,6 @@
 import 'package:extended_image/extended_image.dart';
 import 'package:ff_annotation_route_core/ff_annotation_route_core.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 @FFRoute(
@@ -56,6 +57,27 @@ class _SimplePhotoViewDemoState extends State<SimplePhotoViewDemo> {
               );
             },
           );
+        },
+        // just demo, it the same as default.
+        // if you need to custom it, you can define it base on your case.
+        shouldAccpetHorizontalOrVerticalDrag:
+            (Map<int, VelocityTracker> velocityTrackers) {
+          if (velocityTrackers.keys.length == 1) {
+            return true;
+          }
+
+          // if pointers are not the only, check whether they are in the negative
+          // maybe this is a Horizontal/Vertical zoom
+          Offset offset = const Offset(1, 1);
+          for (final VelocityTracker tracker in velocityTrackers.values) {
+            if (tracker is ExtendedVelocityTracker) {
+              final Offset delta = tracker.getSamplesDelta();
+              offset = Offset(offset.dx * (delta.dx == 0 ? 1 : delta.dx),
+                  offset.dy * (delta.dy == 0 ? 1 : delta.dy));
+            }
+          }
+
+          return !(offset.dx < 0 || offset.dy < 0);
         },
       ),
     );