+++ lib/src/gesture/page_view/gesture_page_view.dart
@@ -48,6 +48,7 @@ class ExtendedImageGesturePageView extends StatefulWidget {
     List<Widget> children = const <Widget>[],
     CanScrollPage? canScrollPage,
     this.preloadPagesCount = 0,
+    this.shouldAccpetHorizontalOrVerticalDrag,
   })  : controller = controller ?? _defaultPageController,
         childrenDelegate = SliverChildListDelegate(children),
         physics = physics != null
@@ -80,6 +81,7 @@ class ExtendedImageGesturePageView extends StatefulWidget {
     int? itemCount,
     CanScrollPage? canScrollPage,
     this.preloadPagesCount = 0,
+    this.shouldAccpetHorizontalOrVerticalDrag,
   })  : controller = controller ?? _defaultPageController,
         childrenDelegate =
             SliverChildBuilderDelegate(itemBuilder, childCount: itemCount),
@@ -102,6 +104,7 @@ class ExtendedImageGesturePageView extends StatefulWidget {
     CanScrollPage? canScrollPage,
     required this.childrenDelegate,
     this.preloadPagesCount = 0,
+    this.shouldAccpetHorizontalOrVerticalDrag,
   })  : controller = controller ?? _defaultPageController,
         physics = _defaultScrollPhysics,
         canScrollPage = canScrollPage ?? _defaultCanScrollPage,
@@ -160,6 +163,11 @@ class ExtendedImageGesturePageView extends StatefulWidget {
 
   /// The count of pre-built pages
   final int preloadPagesCount;
+
+  /// Whether should accpet horizontal or vertical drag at that time
+  /// You can custom it by your base
+  final ShouldAccpetHorizontalOrVerticalDrag?
+      shouldAccpetHorizontalOrVerticalDrag;
 
   @override
   ExtendedImageGesturePageViewState createState() =>
@@ -230,6 +238,8 @@ class ExtendedImageGesturePageViewState
                 () => ExtendedVerticalDragGestureRecognizer(
                   canHorizontalOrVerticalDrag: canHorizontalOrVerticalDrag,
                   debugOwner: this,
+                  shouldAccpetHorizontalOrVerticalDrag:
+                      widget.shouldAccpetHorizontalOrVerticalDrag,
                 ),
                 (ExtendedVerticalDragGestureRecognizer instance) {
                   instance
@@ -254,6 +264,8 @@ class ExtendedImageGesturePageViewState
                 () => ExtendedHorizontalDragGestureRecognizer(
                   canHorizontalOrVerticalDrag: canHorizontalOrVerticalDrag,
                   debugOwner: this,
+                  shouldAccpetHorizontalOrVerticalDrag:
+                      widget.shouldAccpetHorizontalOrVerticalDrag,
                 ),
                 (ExtendedHorizontalDragGestureRecognizer instance) {
                   instance